ref: flex deserves its own folder
