+++ src/layouts/flex/simple.ts
@@ -1,28 +1,35 @@
-import { BoundingBox, Context, FlexStyle, Layout, removeMargins } from "..";
-import { FlexItem } from "./flex-item";
+import {
+  BoundingBox,
+  Context,
+  FlexStyle,
+  Layout,
+  removeMargins
+} from "../../base";
+import { InvalidItemError, ItemWidthError } from "../../errors";
+import { FlexItem } from "./item";
 
-/**
- * Creates a row that draws columns giving them equal spacing
- */
-export class EqualFlex implements Layout {
-  constructor(private style: FlexStyle, private items: FlexItem[]) {}
+export class Flex implements Layout {
+  constructor(private style: FlexStyle, private items: FlexItem[]) {
+    items.forEach(it => {
+      if (!(it instanceof FlexItem)) throw new InvalidItemError();
+      if (!it.itemWidth) throw new ItemWidthError();
+    });
+  }
 
   width(_context: Context, box: BoundingBox) {
     return box.width;
   }
 
   height(context: Context, box: BoundingBox) {
-    // we must factor in the width if the individual items when getting their heigth
-    const width = box.width / this.items.length;
-    return Math.floor(
-      Math.max(...this.items.map(i => i.height(context, { ...box, width })))
+    return Math.max(
+      ...this.items.map(i => i.height(context, { ...box, width: i.itemWidth }))
     );
   }
 
   draw(context: Context, box: BoundingBox) {
     const boxes = this.boxes(context, { ...box });
 
-    this?.style?.background?.draw(context, box);
+    this.style?.background?.draw(context, box);
 
     this.items.forEach((i, c) => {
       i.draw(context, boxes[c]);
@@ -32,14 +39,13 @@ export class EqualFlex implements Layout {
   boxes(context: Context, box: BoundingBox) {
     box = removeMargins(box, this.style.margin);
 
-    const defaultWidth = box.width / this.items.length;
     let originalWidth = 0;
     let floatItemCount = 0;
     let floatSpace = 0;
 
     // check which margins the item has.
     this.items.forEach(i => {
-      originalWidth += defaultWidth;
+      originalWidth += i.itemWidth;
 
       // this item has a left margin
       if (i.flexFloat.includes("left")) floatItemCount++;
@@ -56,7 +62,7 @@ export class EqualFlex implements Layout {
     let x = box.x;
 
     for (const i of this.items) {
-      const width = defaultWidth;
+      const width = i.itemWidth;
       const height = this.height(context, { ...box, width });
 
       // move the box to the left of the FloatSpace
