+++ src/layouts/index.ts
@@ -1,7 +1,4 @@
 export * from "./block";
-export * from "./flex-ratio";
-export * from "./flex-equal";
 export * from "./flex";
-export * from "./flex-item";
 export * from "./padding";
 export * from "./table";
